Style(#233) : 로그인 css - 로그인 인풋 값 placeholeder 컬러 복원
