+++ src/app/(auth)/new-password/page.tsx
@@ -1,5 +1,4 @@
 import FindPassword from '@/components/findpassword/FindPassword'
-import React from 'react'
 
 const NewPasswordPage = () => {
   return <FindPassword />
