+++ src/app/(auth)/signout/page.tsx
@@ -1,11 +1,7 @@
 import LogOutButton from '@/components/logout/LogOutButton'
 
 const signout = () => {
-  return (
-    <div>
-      <LogOutButton />
-    </div>
-  )
+  return <LogOutButton />
 }
 
 export default signout
