+++ src/components/communityDetail/CommunityCreate.tsx
@@ -85,6 +85,7 @@ const CommunityCreate = () => {
 
   if (status === 'unauthenticated') {
     alert('로그인한 유저만 이용 가능합니다.')
+    router.replace('/')
     return
   }
