+++ src/components/login/loginCss.ts
@@ -12,5 +12,8 @@ export const INPUT_FOCUS =
 export const DOWN_ACTIVE_BUTTON =
   'active:text-[rgba(255,255,255,0.3)] active:drop-shadow-none active:shadow-[0px_1px_2px_0px_rgba(0,0,0,0.20)_inset,_0px_4px_4px_0px_rgba(255,255,255,0.20),_0px_0px_0px_1px_rgba(0,0,0,0.10),_0px_0px_0px_4px_rgba(104,91,255,0.10),_0px_0px_0px_3px_rgba(255,255,255,0.20),_0px_4px_1px_0px_rgba(0,0,0,0.20)_inset,_0px_-4px_1px_0px_rgba(255,255,255,0.05)_inset]'
 
+// export const OPEN_ANOTHER_SHADOW =
+//   'shadow-[0px_4px_4px_rgba(0,0,0,0.2),_0px_0px_0px_1px_rgba(0,0,0,0.1),_0px_0px_0px_4px_rgba(0,0,0,0.1),_0px_0px_0px_3px_rgba(255,255,255,0.2),_inset_0px_-4px_1px_rgba(0,0,0,0.1)]'
+
 export const OPEN_ANOTHER_SHADOW =
-  'shadow-[0px_4px_4px_rgba(0,0,0,0.2),_0px_0px_0px_1px_rgba(0,0,0,0.1),_0px_0px_0px_4px_rgba(0,0,0,0.1),_0px_0px_0px_3px_rgba(255,255,255,0.2),_inset_0px_-4px_1px_rgba(0,0,0,0.1)]'
+  'shadow-[0px_4px_4px_rgba(0,0,0,0.2),_0px_0px_0px_1px_rgba(0,0,0,0.1),_0px_0px_0px_4px_rgba(0,0,0,0.1),_0px_0px_0px_3px_rgba(255,255,255,0.2),_inset_0px_1px_2px_rgba(255,255,255,0.2),_inset_0px_4px_1px_rgba(255,255,255,0.2),_inset_0px_-4px_1px_rgba(0,0,0,0.1)]'
